+++ src/components/molecules/index.ts
@@ -1,0 +1,5 @@
+import AddNewItem from './AddNewItem';
+import DataFetcher from './DataFetcher';
+import InitialForm from './InitialForm';
+
+export { AddNewItem, DataFetcher, InitialForm };
